WIP: egl/android: avoid recreating images when we swap buffers
Associate a buffer with an image and when the buffer is dequeued, find
the existing image associated to it.

This avoids recreating the image which involves reallocating a VMA
etc... Potentially generating more stalls down the i915 path as the
kernel needs to serialize work to update the PPGTT tables.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>
Closes: https://gitlab.freedesktop.org/mesa/mesa/issues/2465